requirements.txt
openai
streamlit
streamlit-chat
python-dotenv
beautifulsoup4
requests
langchain
python-docx
reportlab
PyPDF2
faiss-cpu
altair
tiktoken
sentence-transformers
unstructured
